refactor: context, hooks 분리
