+++ src/context/index.ts
@@ -1,0 +1,3 @@
+export * from "./notificationContext";
+export * from "./themeContext";
+export * from "./userContext";
