--- src/hooks/useContext.ts
@@ -1,32 +1,0 @@
-import { useContext } from "react";
-import {
-  NotificationContext,
-  ThemeContext,
-  UserContext,
-} from "../context/appContext";
-
-export const useThemeContext = () => {
-  const context = useContext(ThemeContext);
-  if (context === undefined) {
-    throw new Error("useThemeContext must be used within an ThemeProvider");
-  }
-  return context;
-};
-
-export const useUserContext = () => {
-  const context = useContext(UserContext);
-  if (context === undefined) {
-    throw new Error("useUserContext must be used within an UserProvider");
-  }
-  return context;
-};
-
-export const useNotificationContext = () => {
-  const context = useContext(NotificationContext);
-  if (context === undefined) {
-    throw new Error(
-      "useNotificationContext must be used within an NotificationProvider",
-    );
-  }
-  return context;
-};
