+++ src/hooks/useThemeContext.ts
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import { ThemeContext } from "../context/themeContext";
+
+export const useThemeContext = () => {
+  const context = useContext(ThemeContext);
+  if (context === undefined) {
+    throw new Error("useThemeContext must be used within a ThemeProvider");
+  }
+  return context;
+};
